alexey_project/requirements.txt
asgiref==3.11.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.4
defusedxml==0.7.1
Django==6.0.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
djoser==2.3.3
idna==3.11
inflection==0.5.1
oauthlib==3.3.1
pycparser==3.0
PyJWT==2.11.0
python3-openid==3.2.0
PyYAML==6.0.3
requests==2.32.5
requests-oauthlib==2.0.0
social-auth-app-django==5.7.0
social-auth-core==4.8.3
sqlparse==0.5.5
uritemplate==4.2.0
urllib3==2.6.3
